[BUG] Fix PSR-0 incompatibilty
Having 2 classes in the same file breaks PSR-0 compability
